pyproject.toml
[project]
name = "banner-ai-server"
version = "0.1.0"
description = "banner api server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "boto3>=1.38.32",
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.16.1",
    "langchain[openai]>=0.3.23",
    "langgraph>=0.4.3",
    "langsmith>=0.3.30",
    "onnxruntime>=1.21.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "pytest-playwright>=0.7.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "webcolors>=24.11.1",
    "uvicorn[standard]>=0.29.0",
]
